Light refactors to macro runner

+++ gb_core/include/address_scanner.h
@@ -31,7 +31,7 @@ public:
 	template<typename T>
 	address_scan_result find_value(T target_value) {
 		std::vector<ptrdiff_t> locations;
-		for (int i = 0; i < _size - sizeof(T); i++) {
+		for (size_t i = 0; i < _size - sizeof(T); i++) {
 			if (*((T *) (_memory + i)) == target_value) {
 				locations.push_back(i);
 			}
